#14 Переход на соединение с базой посредством JdbcTemplate. Отображение таблиц доступно на странице tables.jsp добавленным контроллером TablesController. Добавлена текущая конфигурация master-ветки. Внесены изменения в data.sql и schema.sql.

+++ src/main/java/interfaces/Accounts.java
@@ -1,11 +1,12 @@
 package interfaces;
 
+        import java.math.BigDecimal;
         import java.math.BigInteger;
 
 public interface Accounts extends Entity {
     BigInteger getAccountId();
-    BigInteger getPersonId();
+    Integer getPersonId();
     String getCurrency();
-    Double getBalance();
+    BigDecimal getBalance();
     String getDescription();
 }
